+ Add welcome page  + Create onboarding floating div  + Create header component
Relates #17
