+++ pages/onboarding/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+const Heading = styled.h1.attrs(() => ({
+  className: "font-lg font-bold",
+}))``
+
+export { Heading }
